Устранение замечания: Лучше вынести константы в отдельный файл Constants.swift для более удобной навигации по структуре файлов проекта
